Some improvements to the 3D plotting code:
- adaptive sampling in 3D
- better triangle subdivision algorithm (I've named it triforce)
- fix some 'tenting' effects (e.g. a grid square (1,2,3,4) was previously triangulated as (1,2,3), (1,3,4) when (1,2,4), (2,3,4) looks better
- fix some subdivision bugs which were causing gaps in the surfaces
- general code cleanup
- optimise Mesh
- basic (automatic) exclusion support
- implemented `MaxRecursion` option

Examples
--------
(all plotted with `Mesh->All` so the subdivision changes are more visible)
- `Plot3D[x ^ 2 + 1 / y, {x, -1, 1}, {y, 1, 4}]`
Before: 1.8s
![before1](https://cloud.githubusercontent.com/assets/1217643/5439474/307f6902-84d4-11e4-993c-076fb6b67a7f.png)
After: 1.3s
![after1](https://cloud.githubusercontent.com/assets/1217643/5439479/4106f038-84d4-11e4-87db-7d02ae55ed16.png)

- `Plot3D[x y / (x ^ 2 + y ^ 2 + 1), {x, -2, 2}, {y, -2, 2}]`
Before: 2.4s
![before2](https://cloud.githubusercontent.com/assets/1217643/5439475/3871c722-84d4-11e4-8d91-d34f65cab924.png)
After: 0.8s
![after2](https://cloud.githubusercontent.com/assets/1217643/5439482/44ec9298-84d4-11e4-8131-37caa45a1828.png)

- `Plot3D[Sin[x y] /(x y), {x, -3, 3}, {y, -3, 3}]`
Before: 2.3s
![before3](https://cloud.githubusercontent.com/assets/1217643/5439477/3b9a1350-84d4-11e4-9beb-53b350fee8bd.png)
After: 2.7s
![after3](https://cloud.githubusercontent.com/assets/1217643/5439485/4755f8ee-84d4-11e4-8eb0-c4ac62254f89.png)

- `Plot3D[Log[x^2 + y], {x, -1, 1}, {y, -1, 1}]`
Before: 1.9s (This plot is incorrect, it shows the magnitude when `Log` becomes complex valued).
![before4](https://cloud.githubusercontent.com/assets/1217643/5439478/3eda992c-84d4-11e4-8c2b-4698f30b7b11.png)
After: 2.4s
![after4](https://cloud.githubusercontent.com/assets/1217643/5439490/4d2115d8-84d4-11e4-9fce-a5d40ab4100d.png)